# packages/api-client/openapi.json
{"openapi":"3.0.0","paths":{"/api/health":{"get":{"description":"Verifies application runtime and active PostgreSQL database connectivity via Prisma 7.","operationId":"HealthController_getHealth","parameters":[],"responses":{"200":{"description":"Application and database are healthy"},"503":{"description":"Database connectivity failure"}},"summary":"Health check endpoint","tags":["Health"]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User registered successfully with access token"},"400":{"description":"Validation failed or missing fields"},"409":{"description":"Email already registered"},"429":{"description":"Rate limit exceeded"}},"summary":"Register a new user account","tags":["Authentication"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Successfully authenticated, returns access token"},"401":{"description":"Invalid email or password"},"429":{"description":"Rate limit exceeded (5 requests per minute)"}},"summary":"Authenticate user with email and password","tags":["Authentication"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":"Token successfully revoked and invalidated"},"401":{"description":"Missing, invalid, or already revoked access token"}},"security":[{"JWT-auth":[]}],"summary":"Log out user by revoking current JWT token","tags":["Authentication"]}},"/api/auth/me":{"get":{"operationId":"AuthController_getMe","parameters":[],"responses":{"200":{"description":"Authenticated user profile returned"},"401":{"description":"Unauthorized (token missing, invalid, or revoked)"}},"security":[{"JWT-auth":[]}],"summary":"Retrieve currently authenticated user profile","tags":["Authentication"]}},"/api/projects":{"get":{"operationId":"ProjectsController_findAll","parameters":[{"name":"search","required":false,"in":"query","description":"Search term matching project name","schema":{"example":"mobile","type":"string"}},{"name":"status","required":false,"in":"query","description":"Filter by project status","schema":{"type":"string","enum":["NOT_STARTED","IN_PROGRESS","COMPLETED"]}}],"responses":{"200":{"description":"List of owned projects"}},"security":[{"JWT-auth":[]}],"summary":"List all projects belonging to the authenticated user","tags":["Projects"]},"post":{"operationId":"ProjectsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":"Project successfully created"},"400":{"description":"Validation failed or invalid dates"}},"security":[{"JWT-auth":[]}],"summary":"Create a new project for the authenticated user","tags":["Projects"]}},"/api/projects/{id}":{"get":{"operationId":"ProjectsController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"Project UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Project details"},"404":{"description":"Project not found or not owned by user"}},"security":[{"JWT-auth":[]}],"summary":"Get a project by ID (must be owned by user)","tags":["Projects"]},"put":{"operationId":"ProjectsController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Project UUID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":"Project updated successfully"},"400":{"description":"Validation failed or invalid dates"},"404":{"description":"Project not found or not owned by user"}},"security":[{"JWT-auth":[]}],"summary":"Update an existing project owned by user","tags":["Projects"]},"delete":{"operationId":"ProjectsController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Project UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Project deleted successfully"},"404":{"description":"Project not found or not owned by user"}},"security":[{"JWT-auth":[]}],"summary":"Delete a project owned by user","tags":["Projects"]}},"/api/projects/{id}/tasks":{"get":{"operationId":"ProjectsController_findTasks","parameters":[{"name":"id","required":true,"in":"path","description":"Project UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of tasks in project"},"404":{"description":"Project not found or not owned by user"}},"security":[{"JWT-auth":[]}],"summary":"List all tasks for a project owned by user","tags":["Projects"]}},"/api/tasks":{"get":{"operationId":"TasksController_findAll","parameters":[{"name":"search","required":false,"in":"query","description":"Search term matching task name","schema":{"example":"auth","type":"string"}},{"name":"status","required":false,"in":"query","description":"Filter by task status","schema":{"type":"string","enum":["PENDING","IN_PROGRESS","COMPLETED"]}},{"name":"priority","required":false,"in":"query","description":"Filter by task priority","schema":{"type":"string","enum":["LOW","MEDIUM","HIGH"]}},{"name":"projectId","required":false,"in":"query","description":"Filter by project ID","schema":{"example":"123e4567-e89b-12d3-a456-426614174000","type":"string"}}],"responses":{"200":{"description":"List of owned tasks"}},"security":[{"JWT-auth":[]}],"summary":"List all tasks owned by the authenticated user","tags":["Tasks"]},"post":{"operationId":"TasksController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":"Task successfully created"},"400":{"description":"Validation failed"},"404":{"description":"Project not found or not owned by user"}},"security":[{"JWT-auth":[]}],"summary":"Create a new task in a project owned by user","tags":["Tasks"]}},"/api/tasks/{id}":{"get":{"operationId":"TasksController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"Task UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Task details"},"404":{"description":"Task not found or not owned by user"}},"security":[{"JWT-auth":[]}],"summary":"Get task by ID (must belong to project owned by user)","tags":["Tasks"]},"put":{"operationId":"TasksController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Task UUID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"200":{"description":"Task successfully updated"},"400":{"description":"Validation failed"},"404":{"description":"Task not found or not owned by user"}},"security":[{"JWT-auth":[]}],"summary":"Update an existing task owned by user","tags":["Tasks"]},"delete":{"operationId":"TasksController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Task UUID","schema":{"type":"string"}}],"responses":{"200":{"description":"Task successfully deleted"},"404":{"description":"Task not found or not owned by user"}},"security":[{"JWT-auth":[]}],"summary":"Delete a task owned by user","tags":["Tasks"]}},"/api/dashboard":{"get":{"operationId":"DashboardController_getDashboard","parameters":[],"responses":{"200":{"description":"Dashboard metrics calculated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardResponseDto"}}}}},"security":[{"JWT-auth":[]}],"summary":"Get summary metrics for authenticated user","tags":["Dashboard"]}}},"info":{"title":"Project Management System API","description":"Authoritative REST API specification for Web and Mobile clients","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"JWT-auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","description":"Enter JWT access token","in":"header"}},"schemas":{"RegisterDto":{"type":"object","properties":{"fullName":{"type":"string","description":"Full name of the user","example":"Demo User","maxLength":100},"email":{"type":"string","description":"Email address of the user","example":"demo@example.com"},"password":{"type":"string","description":"User password (minimum 8 characters)","example":"SecurePassword123!","minLength":8,"maxLength":128}},"required":["fullName","email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user","example":"demo@example.com"},"password":{"type":"string","description":"User password","example":"SecurePassword123!"}},"required":["email","password"]},"CreateProjectDto":{"type":"object","properties":{"name":{"type":"string","description":"Project name","example":"Mobile App Redesign"},"description":{"type":"string","description":"Project description","example":"Redesign UI components"},"status":{"type":"string","description":"Project status","enum":["NOT_STARTED","IN_PROGRESS","COMPLETED"],"default":"NOT_STARTED"},"startDate":{"type":"string","description":"Start date in ISO format (YYYY-MM-DD)","example":"2026-10-01"},"endDate":{"type":"string","description":"End date in ISO format (YYYY-MM-DD)","example":"2026-12-31"}},"required":["name"]},"UpdateProjectDto":{"type":"object","properties":{"name":{"type":"string","description":"Project name","example":"Mobile App Redesign Updated"},"description":{"type":"string","description":"Project description","example":"Updated description"},"status":{"type":"string","description":"Project status","enum":["NOT_STARTED","IN_PROGRESS","COMPLETED"]},"startDate":{"type":"string","description":"Start date in ISO format (YYYY-MM-DD)","example":"2026-10-01"},"endDate":{"type":"string","description":"End date in ISO format (YYYY-MM-DD)","example":"2026-12-31"}}},"CreateTaskDto":{"type":"object","properties":{"projectId":{"type":"string","description":"Parent Project UUID","example":"123e4567-e89b-12d3-a456-426614174000"},"name":{"type":"string","description":"Task name","example":"Implement auth guard"},"description":{"type":"string","description":"Task description","example":"Extract JWT and verify claims"},"priority":{"type":"string","description":"Task priority","enum":["LOW","MEDIUM","HIGH"],"default":"MEDIUM"},"status":{"type":"string","description":"Task status","enum":["PENDING","IN_PROGRESS","COMPLETED"],"default":"PENDING"},"dueDate":{"type":"string","description":"Due date in ISO format (YYYY-MM-DD)","example":"2026-11-01"}},"required":["projectId","name"]},"UpdateTaskDto":{"type":"object","properties":{"name":{"type":"string","description":"Task name","example":"Updated task name"},"description":{"type":"string","description":"Task description","example":"Updated description"},"priority":{"type":"string","description":"Task priority","enum":["LOW","MEDIUM","HIGH"]},"status":{"type":"string","description":"Task status","enum":["PENDING","IN_PROGRESS","COMPLETED"]},"dueDate":{"type":"string","description":"Due date in ISO format (YYYY-MM-DD)","example":"2026-11-01"}}},"DashboardResponseDto":{"type":"object","properties":{"totalProjects":{"type":"number","description":"Total number of projects owned by user","example":5},"totalTasks":{"type":"number","description":"Total number of tasks across owned projects","example":12},"completedTasks":{"type":"number","description":"Total number of completed tasks","example":7},"pendingTasks":{"type":"number","description":"Total number of pending tasks","example":4},"projectsInProgress":{"type":"number","description":"Total number of projects currently in progress","example":2}},"required":["totalProjects","totalTasks","completedTasks","pendingTasks","projectsInProgress"]}}}}